Add successOrNull and fold extensions
I've written or seen variations of these a few times in the wild, so promoting them to a top-level set of APIs now to ease use.

This introduces three new functional extensions to `ApiResult`: `successOrNull()`, `successOrElse()`, and `fold()`. These allow easy happy path-ing in user code to coerce results into a concrete value.